+++ spec/lib/customized_premailer_spec.rb
@@ -4,15 +4,17 @@ require 'spec_helper'
 
 describe Premailer::Rails::CustomizedPremailer do
   [ :nokogiri, :hpricot ].each do |adapter|
+    next if adapter == :hpricot and RUBY_PLATFORM == 'java'
+
     context "when adapter is #{adapter}" do
       before { Premailer::Adapter.stubs(:use).returns(adapter) }
 
       describe '#to_plain_text' do
         it 'should include the text from the HTML part' do
           premailer =
-            Premailer::Rails::CustomizedPremailer \
+            Premailer::Rails::CustomizedPremailer
               .new(Fixtures::Message::HTML_PART)
-          premailer.to_plain_text.gsub(/\s/, ' ').strip \
+          premailer.to_plain_text.gsub(/\s/, ' ').strip
             .should == Fixtures::Message::TEXT_PART.gsub(/\s/, ' ').strip
         end
       end
@@ -20,12 +22,12 @@ describe Premailer::Rails::CustomizedPremailer do
       describe '#to_inline_css' do
         context 'when inline CSS block present' do
           it 'should return the HTML with the CSS inlined' do
-            Premailer::Rails::CSSHelper \
-              .stubs(:css_for_doc) \
+            Premailer::Rails::CSSHelper
+              .stubs(:css_for_doc)
               .returns('p { color: red; }')
             html = Fixtures::Message::HTML_PART
             premailer = Premailer::Rails::CustomizedPremailer.new(html)
-            premailer.to_inline_css.should include '<p style="color: red;">'
+            premailer.to_inline_css.should =~ /<p style=("|')color: ?red;?\1>/
           end
         end
 
@@ -33,7 +35,7 @@ describe Premailer::Rails::CustomizedPremailer do
           it 'should return the HTML with the CSS inlined' do
             html = Fixtures::Message::HTML_PART_WITH_CSS
             premailer = Premailer::Rails::CustomizedPremailer.new(html)
-            premailer.to_inline_css.should include '<p style="color: red;">'
+            premailer.to_inline_css.should =~ /<p style=("|')color: ?red;?\1>/
           end
         end
       end
@@ -47,13 +49,13 @@ describe Premailer::Rails::CustomizedPremailer do
     end
 
     it 'should pass on the configs' do
-      Premailer::Rails.config = { :foo => :bar }
+      Premailer::Rails.config = { foo: :bar }
       premailer = Premailer::Rails::CustomizedPremailer.new('some html')
       premailer.instance_variable_get(:'@options')[:foo].should == :bar
     end
 
     it 'should not allow to override with_html_string' do
-      Premailer::Rails.config = { :with_html_string => false }
+      Premailer::Rails.config = { with_html_string: false }
       premailer = Premailer::Rails::CustomizedPremailer.new('some html')
       options = premailer.instance_variable_get(:'@options')
       options[:with_html_string].should == true