+++ spec/spec_helper.rb
@@ -1,10 +1,3 @@
-if RUBY_VERSION >= '1.9'
-  require 'coveralls'
-  Coveralls.wear! do
-    add_filter 'spec/'
-  end
-end
-
 require 'premailer/rails'
 
 require 'stubs/action_mailer'
@@ -12,7 +5,7 @@ require 'stubs/rails'
 require 'fixtures/message'
 require 'fixtures/html'
 
-require 'hpricot'
+require 'hpricot' unless RUBY_PLATFORM == 'java'
 require 'nokogiri'
 
 RSpec.configure do |config|